Added Daily Standing

+++ db/migrate/20111226200123_add_rank_to_player.rb
@@ -1,5 +1,5 @@
 class AddRankToPlayer < ActiveRecord::Migration
   def change
-    add_column :players, :rank, :integer
+    add_column :players, :rank, :float
   end
 end
